refactor: extract GitHubExportOptions.swift from GitHubExportConfig.swift
Extracts the `GitHubRepositoryOption` lookup type from
GitHubExportConfig.swift into its own file, mirroring the JiraExportConfig
split (#730). GitHubExportConfig.swift retains only the export
configuration struct.

Byte-identical move of lines 43-74. Both files import Foundation.

GitHubExportConfig.swift: 74 → 41 lines. Tests: 667.
